requirements.txt
boto3
python-docx
docx2pdf
lxml  # This is still required, but pywin32 should be removed
